refactor(web): #2082 — factor out derive-device-name helper from passkey-tile
Move passkey-tile's `deriveDefaultPasskeyName` (UA → "Mac · Safari") to a
shared `lib/auth/derive-device-name.ts` so PR C.2's trusted-browsers list
can render the same labels as the passkey UI without duplicating the
parser. Pure move + import update — no behavior change.
