+++ packages/web/src/ui/components/admin/security/passkey-tile.tsx
@@ -36,40 +36,12 @@ import {
   getPasskeyClient,
   type PasskeyApiError,
 } from "@/lib/auth/passkey-client";
+import { getDefaultDeviceName } from "@/lib/auth/derive-device-name";
 import { useWebAuthnSupported } from "@/ui/hooks/use-webauthn-supported";
 
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
-
-/**
- * Best-effort default name for a fresh passkey, matched off `navigator.userAgent`.
- * The user can always overwrite the field before saving.
- */
-export function deriveDefaultPasskeyName(ua: string): string {
-  const lower = ua.toLowerCase();
-
-  let device = "This device";
-  if (lower.includes("iphone")) device = "iPhone";
-  else if (lower.includes("ipad")) device = "iPad";
-  else if (lower.includes("android")) device = "Android";
-  else if (lower.includes("mac os") || lower.includes("macintosh")) device = "Mac";
-  else if (lower.includes("windows")) device = "Windows PC";
-  else if (lower.includes("linux") || lower.includes("cros")) device = "Linux";
-
-  let browser: string | null = null;
-  if (lower.includes("edg/")) browser = "Edge";
-  else if (lower.includes("chrome/") && !lower.includes("chromium")) browser = "Chrome";
-  else if (lower.includes("firefox/")) browser = "Firefox";
-  else if (lower.includes("safari/") && !lower.includes("chrome/")) browser = "Safari";
-
-  return browser ? `${device} · ${browser}` : device;
-}
-
-function getDefaultName(): string {
-  if (typeof navigator === "undefined") return "This device";
-  return deriveDefaultPasskeyName(navigator.userAgent);
-}
 
 /**
  * Better Auth surfaces user-cancelled WebAuthn flows with code
@@ -154,7 +126,7 @@ export function PasskeyTile({ hasPasskey, onChange }: PasskeyTileProps) {
     }
 
     const id = result.data.id;
-    const defaultName = getDefaultName();
+    const defaultName = getDefaultDeviceName();
     setPending({ id, defaultName });
     setName(defaultName);
   }
